python3-requirements.txt
python-magic-bin
Pillow
pytesseract
python-magic 

capstone
pyqtgraph
pycryptodome
nose
coverage
cstruct
py7zr